점프 거의 다 구현했고 다른거 구현하다 머리 터질거같아서 요기까지

+++ Project_R/Assets/Scripts/Boss/Yasso/YassoStatus.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 public class YassoStatus : MonoBehaviour
 {
     public short iHP = 500;
-    public byte skillUseCount = 0;
-    public byte beenIdleFor = 0;
+    public int skillUseCount = 0;
+    public int beenIdleFor = 0;
 
     public bool bJumping = false;
     public bool bIsDash = false;
